Move logic to count event bookings by role into the database.

+++ src/main/java/uk/ac/cam/cl/dtg/isaac/dto/IsaacEventPageDTO.java
@@ -65,7 +65,7 @@ public class IsaacEventPageDTO extends ContentDTO {
     private Boolean isUserBooked;
     private Boolean isUserOnWaitList;
 
-    private Integer placesAvailable;
+    private Long placesAvailable;
 
 	/**
      *
@@ -434,7 +434,7 @@ public class IsaacEventPageDTO extends ContentDTO {
 	 * getPlacesAvailable based on current bookings..
      * @return the get the places available.
      */
-    public Integer getPlacesAvailable() {
+    public Long getPlacesAvailable() {
         return placesAvailable;
     }
 
@@ -442,7 +442,7 @@ public class IsaacEventPageDTO extends ContentDTO {
 	 * Set the places available based on current bookings.
      * @param placesAvailable - the number of places available.
      */
-    public void setPlacesAvailable(final Integer placesAvailable) {
+    public void setPlacesAvailable(final Long placesAvailable) {
         this.placesAvailable = placesAvailable;
     }
